+++ __tests__/appArticles.test.js
@@ -3,7 +3,6 @@ const testData = require('../db/data/test-data/index');
 const db = require('../db/connection');
 const app = require('../server/app');
 const request = require('supertest');
-const { response } = require('express')
 
 beforeEach(() => seed(testData));
